Rename package controller to edl
