+++ go/edl/utils.go
@@ -12,7 +12,7 @@
    See the License for the specific language governing permissions and
 	 limitations under the License. */
 
-package controller
+package edl
 
 import (
 	"k8s.io/client-go/pkg/api/v1"
